rename ServiceContainer -> OpenChatAgent
