+++ frontend/openchat-client/src/stores/markRead.ts
@@ -4,7 +4,7 @@ import type {
     MarkReadResponse,
     ThreadRead,
     ThreadSyncDetails,
-    ServiceContainer,
+    OpenChatAgent,
 } from "openchat-agent";
 import { unconfirmed } from "./unconfirmed";
 
@@ -82,11 +82,11 @@ export class MessageReadTracker {
         };
     }
 
-    private triggerLoop(api: ServiceContainer): void {
+    private triggerLoop(api: OpenChatAgent): void {
         this.timeout = window.setTimeout(() => this.sendToServer(api), MARK_READ_INTERVAL);
     }
 
-    start(api: ServiceContainer): void {
+    start(api: OpenChatAgent): void {
         if (process.env.NODE_ENV !== "test") {
             this.triggerLoop(api);
         }
@@ -102,7 +102,7 @@ export class MessageReadTracker {
         }
     }
 
-    private sendToServer(api: ServiceContainer): void {
+    private sendToServer(api: OpenChatAgent): void {
         const req = Object.entries(this.state).reduce<MarkReadRequest>((req, [chatId, data]) => {
             if (!data.empty()) {
                 req.push({
@@ -306,6 +306,6 @@ export class MessageReadTracker {
 
 export const messagesRead = new MessageReadTracker();
 
-export function startMessagesReadTracker(api: ServiceContainer): void {
+export function startMessagesReadTracker(api: OpenChatAgent): void {
     messagesRead.start(api);
 }
